requirements.txt
pyrebase4
python-vlc
mutagen
requests
requests_toolbelt
pylint